refactor: keep sorting changes within size budgets
Extract the existing project-tree preferences, sort-runtime cases, and catalog sort helpers so the fix does not increase repository lint baselines.
